--- waterloop-site/src/components/recruitment_form_component/index.tsx
@@ -1,3 +1,0 @@
-import PositionPosting from './PositionPosting'
-
-export default PositionPosting
